==> 01-1-MTA-Data.ipynb <==
import pandas as pd
from tls_paper_funcs import tls_12_appdata_filtering, remove_empty

mta_folder = 'data/mta/'
mta2020 = pd.read_csv(mta_folder + 'joy-tls-ver-suricata.csv', engine='python')

# Our GT:
new_gt = pd.read_csv(mta_folder + 'in-house-gt.csv')

# SSLBLs GT:
def sslbl_to_fp(s):
    return ':'.join(s[i:i+2] for i in range(0, len(s), 2))

sslbl = pd.read_csv(mta_folder + 'sslbl-2022-05-23.csv', engine='python').iloc[::-1].reset_index(drop=True)
sslbl['cert_fp'] = sslbl['SHA1'].map(sslbl_to_fp)
sslbl['comment'] = sslbl['Listingdate'].astype(str).apply(lambda x: 'Listingdate: '+x)
sslbl['new_label'] = sslbl['Listingreason'].apply(lambda x: 'SSLBL '+x)

# Combining the two
groundtruth = pd.concat([new_gt, sslbl[['new_label', 'cert_fp', 'comment']] ], ignore_index=True)
groundtruth['c2'] = 1
groundtruth

#labeled_mta = pd.concat([ mta_2020[relevant_columns], mta_2022[relevant_columns] ]).merge(groundtruth, on='cert_fp', how='left')
labeled_mta = mta2020.merge(groundtruth, on='cert_fp', how='left')
labeled_mta['c2'].fillna(0, inplace=True)
labeled_mta['source'] = 'MTA'
labeled_mta['c2'].value_counts()

labeled_mta['c_tls_version'] = labeled_mta['c_tls_version'].astype(int)
labeled_mta['s_tls_version'] = labeled_mta['s_tls_version'].astype(int)
labeled_mta['c_supported_versions'] = labeled_mta['c_supported_versions'].astype(str).apply(lambda x: x.split('.')[0])
labeled_mta['s_supported_versions'] = labeled_mta['s_supported_versions'].astype(str).apply(lambda x: x.split('.')[0])

tls_version_cols = ['c_tls_version', 's_tls_version', 'c_supported_versions', 's_supported_versions']
labeled_mta[labeled_mta['tls_version'] == 'TLS 1.3'][tls_version_cols].groupby(tls_version_cols).size()

# c_tls_version == s_tls_version = 5, c_supported_versions != (-1, '') s_supported_versions != -1 (== 7f17, fb1a, 304)
# TLS 1.3
labeled_mta[(labeled_mta['c_tls_version']==5)&(labeled_mta['s_tls_version']==5)&
            (labeled_mta['c_supported_versions']!='-1')&(labeled_mta['c_supported_versions']!='')&
            ((labeled_mta['s_supported_versions'] == '7f17')|(labeled_mta['s_supported_versions'] == 'fb1a')|
             (labeled_mta['s_supported_versions'] == '304')|
             (labeled_mta['s_supported_versions'] == 304))]['tls_version'].value_counts()

labeled_mta[~((labeled_mta['c_tls_version']==5)&(labeled_mta['s_tls_version']==5)&
            (labeled_mta['c_supported_versions']!='-1')&(labeled_mta['c_supported_versions']!='')&
            ((labeled_mta['s_supported_versions'] == '7f17')|(labeled_mta['s_supported_versions'] == 'fb1a')|
             (labeled_mta['s_supported_versions'] == '304')|
             (labeled_mta['s_supported_versions'] == 304)))]['tls_version'].value_counts()

labeled_mta['tls_version_guess'] = 'TLS 1.2'
labeled_mta.loc[(labeled_mta['c_tls_version']==5)&(labeled_mta['s_tls_version']==5)&
            (labeled_mta['c_supported_versions']!='-1')&(labeled_mta['c_supported_versions']!='')&
            ((labeled_mta['s_supported_versions'] == '7f17')|(labeled_mta['s_supported_versions'] == 'fb1a')|
             (labeled_mta['s_supported_versions'] == '304')|
             (labeled_mta['s_supported_versions'] == 304)),'tls_version_guess'] = 'TLS 1.3'
labeled_mta['tls_version_guess'].value_counts()

labeled_mta['session_resumed'].value_counts()

labeled_mta[(labeled_mta['session_resumed']==False) & (labeled_mta['tls_version_guess']=='TLS 1.3')]

labeled_mta['tls_dir_19'].value_counts()

# Number of flows that do not reach the 20 record mark:
33030 / (33030 + 6302+937+1)

# Number of flows that were came from our labeling (not SSLBL):
([not x.startswith("SSLBL") for x in labeled_mta['new_label'].value_counts().index.values] * labeled_mta['new_label'].value_counts()).sum()

## MTA Certs

mta2020_certs = pd.read_csv(mta_folder + 'certificates_length_mta.csv', engine='python')
a = mta2020_certs.groupby(['ip.src', 'tcp.srcport', 'ip.dst', 'tcp.dstport', 'fileindex']).size()
for item in a[a>1].index:
    mta2020_certs.drop(mta2020_certs[(mta2020_certs[['ip.src', 'tcp.srcport', 'ip.dst', 'tcp.dstport', 'fileindex']] == item).all(axis=1)].iloc[1:].index, inplace=True)
mta2020_certs

mta2020_certs_merge = mta2020_certs.merge(labeled_mta[labeled_mta['c2']==1][['s_ip','s_port','c_ip','c_port','file', 'c2', 'cert_fp', 'new_label']],
                    left_on=['ip.src', 'tcp.srcport', 'ip.dst', 'tcp.dstport', 'fileindex'], 
                    right_on=['s_ip','s_port','c_ip','c_port','file'], how='left')
mta2020_certs_merge['c2'].fillna(0, inplace=True)
mta2020_certs_merge

## MTA TLS Data

mta_tlsdata = tls_12_appdata_filtering(labeled_mta[labeled_mta['tls_version_guess'] == 'TLS 1.2'])
mta_tlsdata['c2'] = labeled_mta[labeled_mta['tls_version_guess'] == 'TLS 1.2']['c2'].values
remove_empty(mta_tlsdata, True)
display(mta_tlsdata['c2'].value_counts())

id_cols = ['timestamp', 'cert_fp', 'ja3', 'ja3s', 'notafter', 'notbefore', 'subject', 'issuerdn', 'tls_version', 'session_resumed', 'sni_suricata', 'c_ip', 'c_port', 's_ip', 's_port', 'ip_proto', 'start_time', 'c_tls_version', 's_tls_version', 'c_supported_versions', 's_supported_versions',  'file', 'sni_joy', '_merge', 'new_label', 'comment', 'c2', 'source', 'tls_version_guess']
mta_tlsdata[id_cols] = labeled_mta[id_cols]
mta_tlsdata

## Export to CSVs:

#labeled_mta.to_csv('processed-datasets/labeled_mta.csv', index=False)
#mta_tlsdata.to_csv('processed-datasets/mta_tlsdata.csv', index=False)
#mta2020_certs_merge.to_csv('processed-datasets/mta2020_certs_merge.csv', index=False)
