==> 01-4-DoH-Data.ipynb <==
import pandas as pd
import numpy as np
from tls_paper_funcs import tls_12_appdata_filtering, remove_empty, nst_detection_and_removal_after_filtering

This file shows the type of processing that was used to obtain certificate size inference and application behavior extraction from the DoH dataset.  
Unlike others, this notebook has not been converted to a parallel version, and is thus not very efficient.

doh_df = pd.read_csv('data/doh/dataset_patchedjoy2.csv', engine='python')
doh_df

doh_df['tls_dir_19'].value_counts()

586529 /(586529 + 176330 + 61737 + 42)

doh_df['c2'] = 0
doh_df.loc[doh_df['file'].str.startswith('e') |
           doh_df['file'].str.startswith('f') |
           doh_df['file'].str.startswith('g'), 'c2'] = 1

doh_df['tls_version_guess'] = 'TLS 1.2'
doh_df.loc[(doh_df['c_tls_version']==5)&(doh_df['s_tls_version']==5)&
       (doh_df['c_supported_versions']!='-1')&(doh_df['c_supported_versions']!='')&
       ((doh_df['s_supported_versions'] == '7f17')|(doh_df['s_supported_versions'] == 'fb1a')|
       (doh_df['s_supported_versions'] == '304')|(doh_df['s_supported_versions'] == 304)), 'tls_version_guess'] = 'TLS 1.3'

print("Total:",len(doh_df))
print("TLS 1.2: {}, TLS 1.3: {}".format(
            len(doh_df[(doh_df['tls_version_guess'] == 'TLS 1.2')]),
            len(doh_df[(doh_df['tls_version_guess'] == 'TLS 1.3')])
    ))
print("TLS 1.3 no resumptions:", len(doh_df[(doh_df['tls_version_guess'] == 'TLS 1.3') & (doh_df['s_psk'] == -1)]))

display(doh_df['c2'].value_counts())
display(doh_df[doh_df['c2']==1]['tls_version_guess'].value_counts())

# do not perform certificate size guessing here;
# these files already have certificate size guessing:

# from tls_paper_funcs import cert_size_inference_simple_diagram
# tls13samples = list(doh_df[(doh_df['tls_version_guess'] == 'TLS 1.3') & (doh_df['s_psk'] == -1)].index)
# for i in tls13samples:
#    doh_df.loc[i,'cert_size_guess_simple'] = cert_size_inference_simple_diagram(doh_df.loc[i])
#    print(i, end='\r')
# doh_df.to_csv('data/doh/joy-tls-versions-certinf-joypatched.csv', index=False)

# pd.read_csv('data/doh/joy-tls-versions-certinf-joypatched.csv')

# Graph for certificate size distribution:
# _ = doh_df[doh_df['cert_size_guess_simple'] > 0]['cert_size_guess_simple'].hist(bins=100, figsize=(16,6))

tls_tps = ['tls_tp_'+str(x) for x in range(20)]
bad_rows = doh_df[( (doh_df[tls_tps]>23).any(axis=1) |\
                    (((doh_df[tls_tps]<20)&(doh_df[tls_tps]>1)).any(axis=1)) \
                  ) & (doh_df['tls_version_guess'] == 'TLS 1.3') & (doh_df['s_psk'] == -1)].index

# There are rows with errors that should be removed.
doh_df.iloc[bad_rows]

doh13_filtered = tls_12_appdata_filtering( doh_df[ (doh_df['tls_version_guess'] == 'TLS 1.3') \
                                          & (doh_df['s_psk'] == -1) ].drop(bad_rows) ).astype(int)
doh13_filtered[['c_ip', 'c_port', 's_ip', 's_port', 'ip_proto', 'start_time', 'sni', 'file', 'c2']] = \
            doh_df[ (doh_df['tls_version_guess'] == 'TLS 1.3') & (doh_df['s_psk'] == -1) ].drop(bad_rows)[['c_ip', 'c_port', 's_ip', 's_port', 'ip_proto', 'start_time', 'sni', 'file', 'c2']].values
doh13_filtered

remove_empty(doh13_filtered, True)

tls13doh_appdata_filtered = []
tls13samples = list(doh13_filtered.index)
print(len(tls13samples))
for count, i in enumerate(tls13samples):
    nst_discard_b, nst_discard_dir = nst_detection_and_removal_after_filtering(doh13_filtered.loc[i])
    line = np.hstack( [nst_discard_b,   (20 - len(nst_discard_b))*[-1],
                       nst_discard_dir, (20 - len(nst_discard_dir))*[-1]] )
    tls13doh_appdata_filtered.append((i,line))
    print(count, end='\r')

ident_columns = ['c_ip', 'c_port', 's_ip', 's_port', 'ip_proto', 'start_time', 'sni', 'file', 'c2']

tls_bs = ['tls_b_'+str(x) for x in range(20)]
tls_dir = ['tls_dir_'+str(x) for x in range(20)]

doh_tlsdata13 = pd.DataFrame(np.vstack([b for a,b in tls13doh_appdata_filtered]),
                            columns= tls_bs+tls_dir,
                            index=[a for a,b in tls13doh_appdata_filtered])

doh_tlsdata13[ident_columns] = doh13_filtered[ident_columns]

to_remove13 = doh_tlsdata13[ (doh_tlsdata13[tls_bs] == -1).all(axis=1) ].index
print(len(to_remove13))
doh_tlsdata13.drop(to_remove13, inplace=True)

doh_tlsdata13

# doh_tlsdata13.to_csv('data/doh/tls13-behav.csv')
